add libreboot static library
this is a simple library that implements the "reboot_wrapper()"
call, which allows some nifty hax to happen before reboot for
devices that require it, without the need for identical code to
be maintained in multiple locations.

the idea is that ALL PARTS OF ANDROID will use this library to
reboot the system.

Change-Id: Id2c17191ebeb3bb171463f5e6683d6fdf2fff795

fixups for recovery, using __system rather than system

Change-Id: Ia50ab9a0d0e9cd4a62c14ea7ea02fed5470f5ce7

Allow overriding reboot with a default reason if none is provided.

Change-Id: I7ecc979775f5a8852a16693aae10dbcdb81bafa6